populando de forma automatica o DataBase

+++ spring-boot-ionic_nelioAlves/projeto_do_curso/curso_rest/src/main/java/com/projetoNelioAlves/curso_rest/domain/Categoria.java
@@ -15,7 +15,7 @@ public class Categoria implements Serializable {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long id;
+    private Integer id;
 
     private String nome;
 
@@ -23,16 +23,16 @@ public class Categoria implements Serializable {
 
     }
 
-    public Categoria(long id, String nome) {
+    public Categoria(Integer id, String nome) {
         this.id = id;
         this.nome = nome;
     }
 
-    public long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
